Higher order reactions
Dear authors,

to my opinion is the current calculation of the reaction rate is not correct for the case of higher with order > 1. 
For example for a dimerization the current implementation would consider for the reaction: 
Reaction(k,[1,1], [(1,-2), (2,1)])
The rate 
v = k*u[1]*u[1] 
but it should be 
 v = k*u[1]*(u[1]-1)
since a single monomer cannot dimerize with itself. 

The propsed changes in the code should correct this.
I am not sure if it is a nice solution but it works for me. 

Best regrards
DRW