Split +
lololololololol

Okay, real explanation. If you add a ton of things in one line, you get +(...) which makes a giant tuple and then loops through to do the add. But that tuple is large enough that it leaves cache and allocates, since in our example from https://github.com/JuliaDiffEq/OrdinaryDiffEq.jl/issues/664 we have like 255 adds per expression. This adds parenthesis so that way all + operations are pairwise. We might be able to optimize a bit by making it like 8-wise, but on our example it's still 100x faster.