Disable Appveyor nightly testing
Following DiffEqBase and OrdinaryDiffEq. This always fails, and also messes up the build status on the homepage. With all the recent added functionality the test time is approaching 20mins too, so having four sets of tests really slows things down. We can reenable this in the future when a breaking Julia release is approaching (like 2.0).