Process constraint system variables when extending ReactionSystems
Allows `Base.getproperty` to access variables of constraint systems
from `ReactionSystem`s that have been extended. Now, if you `extend` a
`ReactionSystem` with a `Nonlinear` or `ODESystem` variables in this
constraint system will be added to the `var_to_name` dictionary in the
original system.

Two tests, on for variables and one for parameters, are added to test
the new variable access.