Adds the methods MT.defaults, MT.observed and MT.getvar
MT.defaults returns the merged defaults of the ReactionSystem and the
constraint system, if present.

MT.observed returns the merged defaults of the ReactionSystem and the
constraint system, if present.

MT.getvar does the same as ModelingToolkit version, but also checks if
the variable can be found in the constraints system of a
ReactionSystem.

The Base.convert functions for the ReactionSystem to AbstractSystem
conversion have been changed to use the new observed and defaults
methods, so that the observed variables and defaults of constraint
systems are now available from the converted AbstractSystem.

Some regression tests have been added for these new methods.